requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
cattrs==1.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
Faker==6.6.0
idna==2.10
iniconfig==1.1.1
jsonschema==3.2.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-dateutil==2.8.1
requests==2.25.1
schema==0.7.4
six==1.15.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.3
